Refactor asset type validation rule
