+++ source/Energinet.DataHub.MeteringPoints.Application/Create/Consumption/Validation/RuleSet.cs
@@ -24,6 +24,9 @@ namespace Energinet.DataHub.MeteringPoints.Application.Create.Consumption.Valida
     {
         public RuleSet()
         {
+            RuleFor(request => request.AssetType)
+                .SetValidator(new AssetTypeRule()!)
+                .Unless(request => string.IsNullOrWhiteSpace(request.AssetType));
             RuleFor(createMeteringPoint => createMeteringPoint.PowerPlant)
                 .SetValidator(new PowerPlantMustBeValidRule()!)
                 .Unless(request => string.IsNullOrWhiteSpace(request.PowerPlant));
