+++ source/Energinet.DataHub.MeteringPoints.Application/MarketDocuments/ValidationRuleSet.cs
@@ -27,9 +27,8 @@ namespace Energinet.DataHub.MeteringPoints.Application.MarketDocuments
             RuleFor(request => request.ProductType).SetValidator(new ProductTypeRule());
             RuleFor(request => request).SetValidator(new MeasureUnitTypeRule());
             RuleFor(request => request).SetValidator(new MeterReadingOccurenceRule());
-
             RuleFor(request => request).SetValidator(new ActualAddressRule());
-            RuleFor(request => request).SetValidator(new AssetTypeRule());
+
             RuleFor(request => request).SetValidator(new PowerLimitRule());
             RuleFor(request => request).SetValidator(new PhysicalStateRule());
             RuleFor(request => request.TransactionId).SetValidator(new TransactionIdentificationRule());
